python/django/django_by_example/myshop/requirements.txt
celery==3.1.18
Django==1.8.6
django-rosetta==0.7.6
Pillow==3.2.0
pytz==2016.4
WeasyPrint==0.24
django-parler==1.5.1